# Optimize capacity bumping for adaptive ByteBufs