@chrisvest 2025-04-25
Been doing some benchmarking. One thing that shows up is that the histograms have limited fidelity, and the smallest buffer size they can record is 8 KiB. This means this change ends up always allocating 8 KiB for buffers smaller than that, and this ends up using a large amount of memory when the default size is 256 bytes.

The pooled allocator has more granularity in its size classes, and is always able to exact-size the buffers in the `ByteBufAllocatorBenchmark`.

I'll try to see if increasing histogram fidelity helps.

> I have to re-read it better - but since magazine(s) allocations are bump pointer allocations - I was hoping there was the chance of a super fast path where enlarging doesn't cause at all any bytes copy - assuming the current buffer hasn't changed

I worry that this will be too rare to be profitable. Will need a more realistic workload to measure and make sure, though.