@franz1981 2025-04-25
I am not sure is that uncommon; @vietj can confirm: basically there are encoders (of reactive postgres) which doesn't estimate the requirements for the encoded capacity upfront, allocating a single buffer without any capacity set.
Later on, during encoding (spanning several classes which perform it on the event loop), the buffer get enlarged few times (without any further ByteBuf allocation) and, once complete, wrote to the outer layers (likely SSL or via socket).
In such cases, there is not real need of any copies, assuming the super fast path I have explained, to be implemented.