@chrisvest 2025-04-26
<details>
<summary>Current numbers from the ByteBufAllocateAndGrowBenchmark on my M1 Pro</summary>

```
Benchmark                                                              (size)   Mode  Cnt         Score        Error  Units
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree              00256  thrpt   20  22072985.500 ± 540696.942  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgFastCap   00256  thrpt   20       254.901 ±      0.292  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgSize      00256  thrpt   20       254.901 ±      0.292  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree              01024  thrpt   20  22757524.412 ± 675428.307  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgFastCap   01024  thrpt   20       986.720 ±      1.049  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgSize      01024  thrpt   20       594.644 ±      0.632  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree              04096  thrpt   20  15590426.293 ± 783115.792  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgFastCap   04096  thrpt   20      3962.101 ±      5.240  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgSize      04096  thrpt   20      1581.438 ±      2.095  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree              16384  thrpt   20   6723092.103 ± 248097.254  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgFastCap   16384  thrpt   20     16020.518 ±     22.969  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgSize      16384  thrpt   20      4625.471 ±      6.625  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree              65536  thrpt   20   3453546.118 ± 122597.658  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgFastCap   65536  thrpt   20     64256.402 ±     84.232  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgSize      65536  thrpt   20     14474.448 ±     18.976  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree                00256  thrpt   20  16073435.780 ± 229309.257  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgFastCap     00256  thrpt   20       254.838 ±      0.251  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgSize        00256  thrpt   20       254.838 ±      0.251  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree                01024  thrpt   20   5727647.502 ± 421022.762  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgFastCap     01024  thrpt   20       595.003 ±      0.823  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgSize        01024  thrpt   20       595.003 ±      0.823  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree                04096  thrpt   20   2889475.147 ± 166005.486  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgFastCap     04096  thrpt   20      1576.353 ±      4.355  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgSize        04096  thrpt   20      1576.353 ±      4.355  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree                16384  thrpt   20   1292365.419 ±   8228.473  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgFastCap     16384  thrpt   20      4612.601 ±     13.999  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgSize        16384  thrpt   20      4612.601 ±     13.999  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree                65536  thrpt   20    533048.077 ±   2538.240  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgFastCap     65536  thrpt   20     14452.720 ±     49.240  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgSize        65536  thrpt   20     14452.720 ±     49.240  ops/s
```

</details>