@chrisvest 2025-04-26
Marked this as ready for review, but the so-called "super-fast path" is still an interesting idea. Could perhaps be a separate PR.