@chrisvest 2025-04-26
It just occurred to me that change can cause oscillations.

If the allocator correctly predicts the final size of the buffer, it won't get to record that final size in its statistics. It only gets feedback if it predicts too low. So it will predict too low and get feedback about this until the 99-percentile get's high enough, then it predicts correctly for a while, but the buffer size statistics shifts down towards the initial sizes, and eventually the 99-percentile will drop and the allocator will start predicting too low again.

So this needs more work.