@chrisvest 2025-05-07
@franz1981 @normanmaurer This is ready for review. I delayed the collection of buffer size statistics until the time where the pooled `ByteBuf` instances are reused. This slows down the initial warm up a bit, but in turn it means that we only collect statistics about the _final_ size of the buffers, not the initial requested allocations or any capacity bumps.

The benchmarks now look like this, and indicate that this is profitable for all but the smallest buffer sizes:

Before:

```
Benchmark                                                              (size)   Mode  Cnt         Score        Error  Units
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree              00256  thrpt   20  20971200.008 ± 270576.623  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgFastCap   00256  thrpt   20       254.972 ±      0.381  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgSize      00256  thrpt   20       254.972 ±      0.381  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree              01024  thrpt   20   7990032.019 ± 269886.195  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgFastCap   01024  thrpt   20       595.055 ±      0.896  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgSize      01024  thrpt   20       595.055 ±      0.896  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree              04096  thrpt   20   3183902.971 ±  30325.489  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgFastCap   04096  thrpt   20      1579.819 ±      1.505  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgSize      04096  thrpt   20      1579.819 ±      1.505  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree              16384  thrpt   20    680506.519 ±   6548.563  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgFastCap   16384  thrpt   20      4626.348 ±      8.011  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgSize      16384  thrpt   20      4626.348 ±      8.011  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree              65536  thrpt   20    236048.099 ±   2162.341  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgFastCap   65536  thrpt   20     14469.252 ±     15.437  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgSize      65536  thrpt   20     14469.252 ±     15.437  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree                00256  thrpt   20  16120468.646 ± 186072.050  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgFastCap     00256  thrpt   20       254.994 ±      0.345  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgSize        00256  thrpt   20       254.994 ±      0.345  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree                01024  thrpt   20   5174639.204 ±  22944.132  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgFastCap     01024  thrpt   20       594.757 ±      0.665  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgSize        01024  thrpt   20       594.757 ±      0.665  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree                04096  thrpt   20   2726448.069 ±  69778.054  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgFastCap     04096  thrpt   20      1581.071 ±      2.372  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgSize        04096  thrpt   20      1581.071 ±      2.372  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree                16384  thrpt   20   1098746.939 ±  29226.704  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgFastCap     16384  thrpt   20      4624.096 ±      5.646  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgSize        16384  thrpt   20      4624.096 ±      5.646  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree                65536  thrpt   20    492292.545 ±   8670.280  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgFastCap     65536  thrpt   20     14472.056 ±     19.128  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgSize        65536  thrpt   20     14472.056 ±     19.128  ops/s

Benchmark                                                    (size)   Mode  Cnt          Score          Error  Units
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   00064  thrpt   20  201842061.879 ±  7408330.093  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   00256  thrpt   20  202515303.168 ±  7891442.985  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   01024  thrpt   20  205190031.880 ±  7193097.865  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   04096  thrpt   20  208397591.000 ± 16463211.669  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     00064  thrpt   20  103259175.327 ±  2139519.286  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     00256  thrpt   20  153544433.191 ±  3340157.177  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     01024  thrpt   20  100893062.481 ±  5291814.744  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     04096  thrpt   20  122298801.791 ± 16112824.579  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseUnpooled   00064  thrpt    5      22436.659 ±    22835.194  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseUnpooled   00256  thrpt   20    1426013.400 ±   105589.900  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseUnpooled   01024  thrpt   20    1178116.811 ±    57092.008  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseUnpooled   04096  thrpt   20    1005929.958 ±    24847.333  ops/s
```



This PR:

```
Benchmark                                                              (size)   Mode  Cnt         Score        Error  Units
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree              00256  thrpt   20  20343962.801 ± 173329.475  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgFastCap   00256  thrpt   20       254.879 ±      0.254  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgSize      00256  thrpt   20       254.879 ±      0.254  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree              01024  thrpt   20  18976631.103 ± 268202.240  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgFastCap   01024  thrpt   20      1019.946 ±      1.314  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgSize      01024  thrpt   20       594.969 ±      0.767  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree              04096  thrpt   20  15143659.856 ± 288958.291  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgFastCap   04096  thrpt   20      4078.384 ±      4.366  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgSize      04096  thrpt   20      1580.376 ±      1.692  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree              16384  thrpt   20   6061695.690 ±  92124.984  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgFastCap   16384  thrpt   20     16311.035 ±     19.535  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgSize      16384  thrpt   20      4623.908 ±      5.537  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree              65536  thrpt   20   4674994.849 ± 116397.078  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgFastCap   65536  thrpt   20     65229.797 ±    112.870  ops/s
ByteBufAllocateAndGrowBenchmark.adaptiveDirectAllocAndFree:avgSize      65536  thrpt   20     14467.283 ±     25.029  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree                00256  thrpt   20  16230847.487 ± 109882.531  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgFastCap     00256  thrpt   20       254.869 ±      0.318  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgSize        00256  thrpt   20       254.869 ±      0.318  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree                01024  thrpt   20   5247087.457 ±  62515.600  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgFastCap     01024  thrpt   20       595.050 ±      0.808  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgSize        01024  thrpt   20       595.050 ±      0.808  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree                04096  thrpt   20   2894436.400 ±  26465.850  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgFastCap     04096  thrpt   20      1580.243 ±      1.996  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgSize        04096  thrpt   20      1580.243 ±      1.996  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree                16384  thrpt   20   1124120.804 ±  20695.162  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgFastCap     16384  thrpt   20      4623.633 ±      4.167  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgSize        16384  thrpt   20      4623.633 ±      4.167  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree                65536  thrpt   20    494726.460 ±   7822.986  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgFastCap     65536  thrpt   20     14470.926 ±     14.555  ops/s
ByteBufAllocateAndGrowBenchmark.pooledDirectAllocAndFree:avgSize        65536  thrpt   20     14470.926 ±     14.555  ops/s

Benchmark                                                    (size)   Mode  Cnt          Score          Error  Units
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   00064  thrpt   20  248339416.246 ± 14125791.523  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   00256  thrpt   20  260500181.446 ±  2436690.826  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   01024  thrpt   20  227955732.717 ±  9079507.372  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   04096  thrpt   20  257718939.639 ±  1832511.844  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     00064  thrpt   20  156082396.646 ±  1431961.114  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     00256  thrpt   20  128717229.421 ± 20244674.070  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     01024  thrpt   20  123697877.459 ± 17572853.583  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     04096  thrpt   20  127395986.190 ± 27092444.618  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseUnpooled   00064  thrpt    6      16583.195 ±    22116.845  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseUnpooled   00256  thrpt   20    1391611.046 ±   133657.735  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseUnpooled   01024  thrpt   20    1193813.142 ±    59223.684  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseUnpooled   04096  thrpt   20    1027168.954 ±    23529.722  ops/s
```